blankon/requirements.txt
channels==4.1.0
channels-redis==4.2.0
django==4.2.11
djangorestframework==3.15.1
drf-yasg==1.21.5
djangorestframework-simplejwt==4.6.0
daphne==4.1.2
psycopg2==2.9.2